refactor(dynacell): assorted small simplifications
- `metrics.py:features_from_crops` — the per-cell fallback returned whatever
  dtype the extractor produced while the batch path cast to float32, so the two
  paths could write different dtypes into the same feature cache. Cast both.
- `pipeline_cache.py` — flip the `pending` conditional so the force-recompute
  branch reads first-to-last.
- `reporting/tables.py:aggregate_metrics` — drop an `agg = ...; return agg`.
- `preprocess/zarr_utils.py:rewrite_zarr` — walked `positions()` twice (once to
  count for tqdm, once to iterate); materialize the list once.

+++ applications/dynacell/src/dynacell/evaluation/metrics.py
@@ -722,7 +722,8 @@ def features_from_crops(crops, feature_extractor):
         out = batch_fn(crops)
         return np.asarray(out.detach().cpu()).reshape(len(crops), -1).astype(np.float32, copy=False)
     feats = [feature_extractor.extract_features(c).detach().cpu().numpy().reshape(-1) for c in crops]
-    return np.stack(feats, axis=0)
+    # float32 to match the batch path, so both write the same dtype to the cache.
+    return np.stack(feats, axis=0).astype(np.float32, copy=False)
 
 
 def build_crops(image, cell_segmentation, patch_size, *, z_slab: slice | None = None):
